Created Data class. Tested string.
Not using an existing timer library for Arduino because those require callback functions with signature (void *) f(), which doesn't work with a Data class. May rework into struct and associated modifying functions to see if modifying a global struct is an acceptable callback for those libraries. But the timing implemented in update_continuous_readings (without interrupts) seems to be working OK.
TODO: figure out why WiFiSSLClient seems to fail to connect to any servers... Or otherwise figure out how to send the String over WiFi.